Grid and position changes
- changed and reset the transform positions on most game objects so they
conform to regular units for a grid
- each square is 1 by 1 unit and is lined up so that the origin is on a
line. This allows for easier positioning as we now only have to deal
with whole numbers or halves when positioning things.
- Arrows now snap to a grid (they favor the left if there's any doubt,
I'll fix that later)
- Rover now snaps to an arrow when it needs to change direction (it
looks bad, will fix later)
- Arrows start out in the lower left
- Changed text to tell you that you need to press space

Just a note that if you make a new gameobject to act as a parent for
other things, reset the transform so that it doesn't add the position to
its children without explicitly adding it.