Added music and sounds
Any and all sound can be turned off for debugging in the Gamecontroller
object. Uncheck Play Music to make only sound effects play, and uncheck
game sound to turn off all sounds. Ideally we'll put this into an
options menu in the final game and this will make it easier.

-Added background music "Acid Rain"
-Added sound effects for rover engine, dropping an arrow, and picking up
a mineral