Fixed Collision and Level 3
- Stone blocks now behave as intended and ignore arrows through
different collision layers. This changed some of the components of other
game objects but they should all behave identically to what they did
before.
-Minerals out is automatically calculated so there shouldn't be any
coding problems with adding fewer or more minerals to the levels
- added a few more tags to make things simpler. anything that is
pick-up-able should be tagged "Pick-up" and any arrows/commands should
be tagged "arrow" and should be on the "Commands" layer
-Cleaned up a bit of the code on the building script
- Level 3 is now possible